Backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# HTTP client for NetSuite API calls
httpx>=0.24.0
requests>=2.28.0
requests-oauthlib>=1.3.0

# Data validation and serialization
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment variables management
python-dotenv>=1.0.0

# Date and time handling
python-dateutil>=2.8.0

# JSON handling and data processing
orjson>=3.8.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Development and testing (optional)
#pytest>=7.0.0
#pytest-asyncio>=0.20.0
#black>=23.0.0
#flake8>=6.0.0

# CORS support for frontend integration
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
pydantic[email]

# Logging
structlog>=23.0.0

uvloop>=0.16.0
httptools>=0.5.0

# For background CSV writer (asyncio queue)
aiofiles>=23.0.0